requirements.txt
# Flask Framework
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-Login>=0.6.3
Werkzeug>=3.0.0

# MongoDB & Database
pymongo>=4.5.0
python-dotenv>=1.0.0

# Authentication & Security
itsdangerous>=2.1.0

# AI/ML
google-genai>=0.3.0

# Additional utilities (if needed in future)
# requests>=2.31.0
# python-dateutil>=2.8.2